analyzer/requirements.txt
connexion[flask]
connexion[uvicorn]
swagger-ui-bundle
pyyaml
kafka-python
starlette